Final Commit.
Final commit after Colorado Regional. We got to the semis, but were
crushed. Code is a bit of a mess because of all the Autonomous changes
that were made... I may clean it up later, I haven't decided if it's
worth it. Over the summer we may work on getting a perfect Auto as an
offseason activity.
-Isaac